moved info to chat
