+++ resources/views/livewire/chat/drawer.blade.php
@@ -1,4 +1,4 @@
-<div >
+<div>
 
     @script
         <script>
