+++ src/Livewire/Chat/Group/AddMembers.php
@@ -8,7 +8,7 @@ use Livewire\Attributes\Validate;
 //use Livewire\Features\SupportFileUploads\TemporaryUploadedFile;
 use Livewire\WithFileUploads;
 use Namu\WireChat\Facades\WireChat;
-use Namu\WireChat\Livewire\Info\Info;
+use Namu\WireChat\Livewire\Chat\Info;
 use Namu\WireChat\Livewire\Modals\ModalComponent;
 use Namu\WireChat\Models\Conversation;
 use Namu\WireChat\Models\Participant;
