+++ src/Livewire/Chat/Group/Members.php
@@ -13,7 +13,7 @@ use Namu\WireChat\Enums\ParticipantRole;
 use Namu\WireChat\Facades\WireChat;
 use Namu\WireChat\Livewire\Modals\ModalComponent;
 use Namu\WireChat\Livewire\Components\WireChat as WidgetsWireChat;
-use Namu\WireChat\Livewire\Info\Info;
+use Namu\WireChat\Livewire\Chat\Info;
 use Namu\WireChat\Models\Action;
 use Namu\WireChat\Models\Conversation;
 use Namu\WireChat\Models\Participant;
